Moved components out of rotes folder.
